+++ modules/hanna-react/src/MainMenu2.tsx
@@ -463,19 +463,20 @@ export const MainMenu2 = (props: MainMenu2Props) => {
             type="button"
             aria-pressed={isMenuOpen}
             aria-controls={menuId}
-            data-icon="text"
             {...(isMenuOpen
               ? {
                   onClick: closeMenu,
                   'aria-label': txt.closeMenuLong,
                   title: txt.closeMenuLong,
                   children: txt.closeMenu,
+                  'data-icon': 'text',
                 }
               : {
                   onClick: openMenu,
                   'aria-label': txt.openMenuLong,
                   title: txt.openMenuLong,
                   children: txt.openMenu,
+                  'data-icon': 'close',
                 })}
           />
         ) : (